requirements.txt
asgiref==3.5.2
Django==3.2
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
mysqlclient==2.1.1
pytz==2022.1
sqlparse==0.4.2
